requirements.txt
slixmpp>=1.8.3
aiofiles>=23.2.1
python-dotenv>=1.0.0
httpx>=0.24.1
pydantic>=2.0.0
asyncio>=3.4.3
rich>=13.5.2
python-dateutil>=2.8.2